pyproject.toml
[project]
name = "crypto-tracker-api"
version = "0.1.0"
description = ""
authors = [
    {name = "AlexMaysonJr",email = "alex.mayson.jr@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.16.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "crypto_tracker_api", from = "src"}]
package-mode = false


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
